fixed some float casts

+++ src/core/soloud_misc.cpp
@@ -53,7 +53,7 @@ namespace SoLoud
 					float f = 0;
 					for (int i = 1; i < 22; i += 2)
 					{
-						f += (4.0f / (M_PI * i)) * (float)sin(2 * M_PI * i * p);
+						f += (float)(4.0f / (M_PI * i)) * (float)sin(2 * M_PI * i * p);
 					}
 					return f * 0.5f;
 				}
@@ -63,9 +63,9 @@ namespace SoLoud
 					for (int i = 1; i < 15; i++)
 					{
 						if (i & 1)
-							f += (1.0f / (M_PI * i)) * sin(p * 2 * M_PI * i);
+							f += (float)(1.0f / (M_PI * i)) * (float)sin(p * 2 * M_PI * i);
 						else
-							f -= (1.0f / (M_PI * i)) * sin(p * 2 * M_PI * i);
+							f -= (float)(1.0f / (M_PI * i)) * (float)sin(p * 2 * M_PI * i);
 					}
 					return f;
 				}
